Split ApplicationHelper into focused helper modules
ApplicationHelper had grown to ~22 methods spanning unrelated concerns. Extract
them into dedicated, cohesive helpers, leaving ApplicationHelper with only
app/layout/hash utilities:

- SortingHelper    - sort_link (+ SORT_NAMES)
- FormattingHelper - object_value_to_s, format_memory, number_with_delimiter, truncate
- BadgesHelper     - status_badge, kafka_state_badge, lag_trend_badge, tags
- PartitionsHelper - partition_replica_brokers/in_sync_brokers, lso_risk_state_bg/badge,
                     offset_with_label, lag_with_label, normalized_metric

All modules are included into Ui::Base (inherited by both OSS and Pro apps), so
view call sites are unchanged. Existing specs move to matching *_helper_test.rb
files.
